+++ yash-env/CHANGELOG.md
@@ -56,8 +56,8 @@ A _private dependency_ is used internally and not visible to downstream users.
     - `system::Sigmask`
     - `system::Wait`
     - `trap::SignalSystem`
-- The `trap::Condition::to_string` method now requires a `system::Signals`
-  implementation instead of `trap::SignalSystem`.
+- The `trap::Condition::to_string` and `trap::Condition::iter` methods now
+  require a `system::Signals` implementation instead of `trap::SignalSystem`.
 - `<subshell::Subshell as std::fmt::Debug>::fmt` now includes the `job_control`
   and `ignores_sigint_sigquit` fields in its output.